# sms/service/server_service/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for VCS System Management API.",
        "title": "VCS System Management API",
        "contact": {
            "name": "Vu Dang Duc Minh",
            "url": "https://github.com/vudangducminh",
            "email": "vudangducminh@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8801",
    "basePath": "/api/v1",
    "paths": {
        "/server/add_server": {
            "post": {
                "description": "Add a new server by validating input and storing server information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Add a new server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token for authentication",
                        "name": "jwt",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add server request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.AddServerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Server added",
                        "schema": {
                            "$ref": "#/definitions/entities.AddServerSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/entities.AddServerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed",
                        "schema": {
                            "$ref": "#/definitions/entities.AuthErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Server already exists",
                        "schema": {
                            "$ref": "#/definitions/entities.AddServerConflictResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/entities.AddServerInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/server/delete_server": {
            "delete": {
                "description": "Delete a server by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Delete a server by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token for authentication",
                        "name": "jwt",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Delete server request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.DeleteServerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/entities.DeleteServerResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed",
                        "schema": {
                            "$ref": "#/definitions/entities.AuthErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Server not found",
                        "schema": {
                            "$ref": "#/definitions/entities.DeleteServerStatusNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/entities.DeleteServerInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/server/export_excel": {
            "get": {
                "description": "Export server data to an Excel file with optional filtering and ordering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Export server data to Excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token for authentication",
                        "name": "jwt",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Order of results, either 'asc' or 'desc'. If not provided or using the wrong order format, the default order is ascending",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by server_name, ipv4, or status. If not provided or using the wrong filter format, then there is no filter applied",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Substring to search in server_name, ipv4, or status",
                        "name": "string",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel file exported successfully",
                        "schema": {
                            "$ref": "#/definitions/entities.ExportExcelSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/entities.ExportExcelBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed",
                        "schema": {
                            "$ref": "#/definitions/entities.AuthErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to export into Excel file",
                        "schema": {
                            "$ref": "#/definitions/entities.ExportExcelFailedResponse"
                        }
                    }
                }
            }
        },
        "/server/import_excel": {
            "post": {
                "description": "Import server data from an Excel file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Import file from excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token for authentication",
                        "name": "jwt",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Excel file to import",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel file imported successfully",
                        "schema": {
                            "$ref": "#/definitions/entities.ImportExcelSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Failed to retrieve file",
                        "schema": {
                            "$ref": "#/definitions/entities.ImportExcelRetrieveFailedResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed",
                        "schema": {
                            "$ref": "#/definitions/entities.AuthErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to add server to Elasticsearch from Excel row",
                        "schema": {
                            "$ref": "#/definitions/entities.ImportExcelElasticsearchErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.AddServerBadRequestResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
        "entities.AddServerConflictResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Server already exists"
                }
            }
        },
        "entities.AddServerInternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
        "entities.AddServerRequest": {
            "type": "object",
            "properties": {
                "ipv4": {
                    "description": "IPv4 address of the server",
                    "type": "string"
                },
                "server_name": {
                    "type": "string"
                },
                "status": {
                    "description": "e.g., \"active\", \"inactive\", \"maintenance\"",
                    "type": "string"
                }
            }
        },
        "entities.AddServerSuccessResponse": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "Unique identifier for the server",
                    "type": "string"
                },
                "ipv4": {
                    "description": "IPv4 address of the server",
                    "type": "string"
                },
                "message": {
                    "description": "Confirmation message",
                    "type": "string",
                    "example": "Server added"
                },
                "server_name": {
                    "description": "Name of the server",
                    "type": "string"
                }
            }
        },
        "entities.AuthErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Authentication failed"
                }
            }
        },
        "entities.DeleteServerInternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
        "entities.DeleteServerRequest": {
            "type": "object",
            "required": [
                "_id"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                }
            }
        },
        "entities.DeleteServerResponse": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "Unique identifier for the deleted server",
                    "type": "string"
                },
                "ipv4": {
                    "description": "IPv4 address of the deleted server",
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Server deleted successfully"
                },
                "server_name": {
                    "description": "Name of the deleted server",
                    "type": "string"
                }
            }
        },
        "entities.DeleteServerStatusNotFoundResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Server not found"
                }
            }
        },
        "entities.ExportExcelBadRequestResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request parameters"
                }
            }
        },
        "entities.ExportExcelFailedResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Failed to export into Excel file"
                }
            }
        },
...
        "entities.ExportExcelSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Excel file exported successfully"
                }
            }
        },
        "entities.ImportExcelElasticsearchErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Failed to add server to Elasticsearch from Excel row"
                }
            }
        },
...
        "entities.ImportExcelRetrieveFailedResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Failed to retrieve file"
                }
            }
        },
        "entities.ImportExcelSuccessResponse": {
            "type": "object",
            "properties": {
                "added_servers": {
                    "type": "integer",
                    "example": 5
                },
                "error_servers": {
                    "type": "integer",
                    "example": 2
                },
                "message": {
                    "type": "string",
                    "example": "Excel file imported successfully"
                }
            }
        },
...
    },
...
}